requirement.txt
PyQt5==5.15.9
vtk==9.2.6
pydicom==2.4.1
Pillow==9.5.0
numpy==1.24.3